Completed CSES's Searching_Sorting

+++ cp/Rawtemplates/Binary_Expo_with_MOD.cpp
@@ -1,6 +1,6 @@
-long long binpow(long long a, long long b, long long m) {
+int binpow(int a, int b, int m) {
     a %= m;
-    long long res = 1;
+    int res = 1;
     while (b > 0) {
         if (b & 1)
             res = res * a % m;
